split benchmark and systems under test into separate files
